Ray work space

+++ BackEnd/QArte/QArte.Services/Services/UserService.cs
@@ -76,9 +76,12 @@ namespace QArte.Services.Services
                 await _pageService.TotalDeleteAsync(page.ID);
             }
 
-            var region = RegionEndpoint.EUCentral1;
-            AmazonS3Client client = new AmazonS3Client(_amazonData.AccessKey, _amazonData.SecretKey, region);
-            await client.DeleteObjectAsync(_amazonData.BucketName, user.PictureUrl);
+            if(user.PictureUrl!= "Public_Resources/QArte_BO.jpg")
+            {
+                var region = RegionEndpoint.EUCentral1;
+                AmazonS3Client client = new AmazonS3Client(_amazonData.AccessKey, _amazonData.SecretKey, region);
+                await client.DeleteObjectAsync(_amazonData.BucketName, user.PictureUrl);
+            }
 
             _qarteDBContext.Users.Remove(user);
 
@@ -341,7 +344,7 @@ namespace QArte.Services.Services
         public async Task<UserDTO> PostAsync(UserDTO obj)
         {
             //make it fetch the default picture
-            string defaultPicPath = "Public_Resources/QArte_B.png";
+            string defaultPicPath = "Public_Resources/QArte_BO.jpg";
 
             obj.PictureURL = defaultPicPath;
